Updated script.js & style.css files
Fixed the toggle which now works and changes theme from dark to light and back.
Updated app colours to be pretty much WCAG compliant.
Still need to update the code, so that it reads browser's settings and if the dark mode is enabled, it loads the app with the toggle switch moved to the right.